test: 유닛 & 슬라이스 테스트 제거
