+++ src/test/java/roomescape/controller/UserWebControllerTest.java
@@ -1,34 +1,40 @@
 package roomescape.controller;
 
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.test.annotation.DirtiesContext;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@WebMvcTest(UserWebController.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 class UserWebControllerTest {
 
-    @Autowired
-    private MockMvc mockMvc;
+    @LocalServerPort
+    int randomServerPort;
+
+    @BeforeEach
+    public void setup() {
+        RestAssured.port = randomServerPort;
+    }
 
     @DisplayName("/로 요청하면 200응답이 넘어온다.")
     @Test
-    void requestAdminPageTest() throws Exception {
-        mockMvc.perform(get("/"))
-                .andDo(print())
-                .andExpect(status().isOk());
+    void requestPopularThemePageTest() {
+        RestAssured.given().log().all()
+                .when().get("/")
+                .then().log().all()
+                .statusCode(200);
     }
 
     @DisplayName("/reservation으로 요청하면 200응답이 넘어온다.")
     @Test
-    void requestReservationPageTest() throws Exception {
-        mockMvc.perform(get("/reservation"))
-                .andDo(print())
-                .andExpect(status().isOk());
+    void requestUserReservationPageTest() {
+        RestAssured.given().log().all()
+                .when().get("/reservation")
+                .then().log().all()
+                .statusCode(200);
     }
 }
