+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -21,7 +21,7 @@ import static org.springframework.test.context.jdbc.Sql.ExecutionPhase.BEFORE_TE
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @Sql(value = {"/schema.sql", "/data.sql"}, executionPhase = BEFORE_TEST_METHOD)
-class ReservationTimeServiceIntegrationTest {
+class ReservationTimeServiceTest {
 
     @Autowired
     private ReservationTimeService reservationTimeService;
